refactor: money패키지명 totalmoney로 변경
