+++ src/test/java/model/totalmoney/TotalPurchaseMoneyTest.java
@@ -1,4 +1,4 @@
-package model.money;
+package model.totalmoney;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
